+++ src/components/Competition/competition.module.css
