+++ src/components/Competition/index.ts
@@ -1,0 +1,1 @@
+export { Competition } from "./Competition";
